[Feat] 탱커, 원거리딜러, 근접 딜러, 서포터, 숲의 순찰자 시너지 제작

+++ Assets/Workspace/HSD/Scripts/Unit/Buff/IEffectable.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 
 public interface IEffectable
 {
-    void ApplyEffect(BuffEffectData buffEffectData, int value, string source);
+    void ApplyEffect(BuffEffectData buffEffectData, float value, string source);
 }
